+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Optionally break strings for `ellipsis()` at a specific character.
 
 ## [2.0.0] - 2022-10-12
 ### Added